Alterações finais

+++ GerenciadorVeiculos/Navio.cs
@@ -8,7 +8,7 @@ namespace GerenciadorVeiculos
 {
     abstract class Navio : Veiculo
     {
-        protected Navio(string id, Modelo modelo, int veloAtual, int peso, int passageiros) : base(id, modelo, veloAtual, peso, passageiros)
+        protected Navio(string id, Modelo modelo, int veloAtual, double peso, int passageiros) : base(id, modelo, veloAtual, peso, passageiros)
         {
         }
