+++ GerenciadorVeiculos/NavioPasseio.cs
@@ -8,7 +8,7 @@ namespace GerenciadorVeiculos
 {
     internal class NavioPasseio : Navio
     {
-        public NavioPasseio(string id, Modelo modelo, int veloAtual, int peso, int passageiros) : base(id, modelo, veloAtual, peso, passageiros)
+        public NavioPasseio(string id, Modelo modelo, int veloAtual, double peso, int passageiros) : base(id, modelo, veloAtual, peso, passageiros)
         {
         }
     }
